+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - New #199: Add `guestMessage` field to `Update` type.
 - New #199: Add `SentGuestMessage` type.
 - New #199: Add `AnswerGuestQuery` method.
+- New #199: Add `canReactToMessages` field to `ChatMemberRestricted` and `ChatPermissions` types.
 
 ## 0.20 April 26, 2026